api response structure improved + inventory history page added

+++ backend/src/modules/cart/cart.controller.ts
@@ -1,18 +1,22 @@
 import { Request, Response } from "express";
 import { Cart } from "./cart.model.js";
+import logger from "../../config/logger.config.js";
 
 export const getCart = async (req: Request, res: Response) => {
   const userId = req.user?.id;
   try {
     const cart = await Cart.find({ userId: userId });
     if (!cart) {
-      res.status(404).json({ message: "Cart not found" });
+      res
+        .status(404)
+        .json({ message: "Cart not found", error: "Cart not found" });
       return;
     }
     res.status(200).json({ message: "Successful", data: cart });
     return;
   } catch (error: any) {
-    res.status(500).json({ error: error.message });
+    logger.error(error);
+    res.status(500).json({ message: "Unsuccessful", error: error.message });
     return;
   }
 };
@@ -44,42 +48,69 @@ export const addToCart = async (req: Request, res: Response) => {
     res.status(201).json({ message: "Successful", data: cart });
     return;
   } catch (error: any) {
-    res.status(500).json({ error: error.message });
+    logger.error(error);
+    res.status(500).json({ message: "Unsuccessful", error: error.message });
     return;
   }
 };
 
 export const updateCartItem = async (req: Request, res: Response) => {
-  const { userId, cartId } = req.params;
+  const userId = req.user?.id;
+  const { cartId } = req.params;
   const { quantity } = req.body;
 
   try {
     const cart = await Cart.findById(cartId);
     if (!cart) {
-      res.status(404).json({ message: "Cart not found" });
+      res
+        .status(404)
+        .json({ message: "Cart not found", error: "Cart not found" });
+      return;
+    }
+    if (cart.userId.toString() != userId) {
+      res
+        .status(403)
+        .json({
+          message: "Invalid credentials",
+          error: "Cart owner is not same as authorized user",
+        });
       return;
     }
     cart.quantity = quantity;
     await cart.save();
     res.status(200).json({ message: "Successful", data: cart });
   } catch (error: any) {
-    res.status(500).json({ error: error.message });
+    logger.error(error);
+    res.status(500).json({ message: "Unsuccessful", error: error.message });
   }
 };
 
 export const deleteCartItem = async (req: Request, res: Response) => {
-  const { userId, cartId } = req.params;
-
+  const { cartId } = req.params;
+  const userId = req.user?.id;
   try {
-    const cart = await Cart.findByIdAndDelete(cartId);
+    const cart = await Cart.findById(cartId);
     if (!cart) {
-      res.status(404).json({ message: "Cart not found" });
+      res
+        .status(404)
+        .json({ message: "Cart not found", error: "Cart not found" });
       return;
     }
-    res.status(200).json({ message: "Product deleted successfully" });
+    if (cart.userId.toString() !== userId) {
+      res
+        .status(403)
+        .json({
+          message: "Invalid credentials",
+          error: "Cart owner is not same as authorized user",
+        });
+      return;
+    }
+    await Cart.findByIdAndDelete(cartId);
+    res.status(200).json({ message: "Successful" });
     return;
   } catch (error: any) {
-    res.status(500).json({ error: error.message });
+    logger.error(error);
+    res.status(500).json({ message: "Unsuccessful", error: error.message });
     return;
   }
 };
